fixed aiming error
The way I converted the coordinates that SDL provided me to coordinates that Ogre wanted to use was not working correctly. I fixed this and also decided to do the conversion just once when collecting the input, so that was every time I need to use it I dont have to re-run the conversion.

+++ include/InputState.h
@@ -7,8 +7,8 @@ struct InputState
     bool left = false;
     bool back = false;
     bool right = false;
-    int mLocationX = 0;
-    int mLocationY = 0;
+    float mLocationX = 0.0f;
+    float mLocationY = 0.0f;
     bool primarySpellPressed = false;
     bool secondarySpellPressed = false;
     bool primarySpellReleased = false;
